+++ examples/observer_node/MyMesh.h
@@ -15,7 +15,7 @@
 #endif
 
 #ifndef FIRMWARE_VERSION
-#define FIRMWARE_VERSION "v1.16.0-observer"
+#define FIRMWARE_VERSION "v1.17.1-observer"
 #endif
 
 #if defined(NRF52_PLATFORM) || defined(STM32_PLATFORM)